requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.12.47
botocore==1.15.47
chardet==3.0.4
docutils==0.15.2
idna==2.9
jmespath==0.9.5
multidict==4.7.5
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
slackclient==2.5.0
urllib3==1.25.9
yarl==1.4.2
